Create rudimentary bulk import script
Based on the testing script (suggested in issue #88), the bulk_import.py script will iterate over each of the parts in bulk_import.yaml and import them into Inventree using data from Digi-Key.

The format and syntax of the yaml file is the same as the testing yaml, except each of the parts are under a single "Parts" List (though they are correctly imported into Digi-key under the appropriate category).

Now if only there was a way to import it with a number in stock as well...